+++ web-sync/Repositories/ob/CountryObRepository.cs
@@ -1,18 +1,19 @@
 ﻿using Dapper;
 using Npgsql;
-using sync_data.Dtos;
-using sync_data.Models.ob;
+using web_sync.Dtos;
+using web_sync.Models.ob;
 
-namespace sync_data.Repositories.ob
+namespace web_sync.Repositories.ob
 {
     public interface ICountryObRepository
     {
-        IEnumerable<CountryObModel> GetAll(CountryDto param);
+        Task<IEnumerable<CountryObModel>?> GetAll(CountryDto param);
+        Task<CountryObModel?> GetById(long id);
         void Insert(CountryObModel Country);
         void ReplaceInto(CountryObModel Country);
         void BulkInsert(List<CountryObModel> Country);
-        void Update(int id, CountryObModel Country);
-        void Delete(int id);
+        void Update(long id, CountryObModel Country);
+        void Delete(long id);
     }
 
     public class CountryObRepository : ICountryObRepository
@@ -25,16 +26,21 @@ namespace sync_data.Repositories.ob
             _connection = connection;
         }
 
-        public IEnumerable<CountryObModel> GetAll(CountryDto param)
+        public async Task<IEnumerable<CountryObModel>?> GetAll(CountryDto param)
         {
             string fields = "*";
             string where = " WHERE true ";
             string limit = "";
             string sort = "";
 
-            if(param.CountryId != null)
+            if (param.CountryId != null)
             {
                 where += " AND country_id = @CountryId";
+            }
+
+            if (param.FromCountryId != null)
+            {
+                where += " AND country_id > @FromCountryId";
             }
 
             if (param.CountryCode != null)
@@ -53,29 +59,26 @@ namespace sync_data.Repositories.ob
                 where += " AND region_id = @RegionId";
             }
 
+            if (param.Offset != null)
+            {
+                limit += " OFFSET " + param.Offset.ToString();
+            }
 
             if (param.Limit != null)
             {
-                if(param.Offset != null)
-                {
-                    limit += " limit " + param.Offset.ToString() + ", " + param.Limit.ToString();
-                }
-                else
-                {
-                    limit += " limit " + param.Limit.ToString();
-                }     
+                limit += " LIMIT " + param.Limit.ToString();
             }
 
-            if(param.SortBy != null)
+            if (param.SortBy != null)
             {
-                string sortOrder = param.SortOrder != "asc" ? " desc": " asc" ;
+                string sortOrder = param.SortOrder != "asc" ? " desc" : " asc";
                 string[] sortBy = { "country_id", "region_id", "country_code", "country_name" };
-                sort += " ORDER BY " + (sortBy.Contains(param.SortBy) ? param.SortBy: sortBy[0]) + sortOrder;
+                sort += " ORDER BY " + (sortBy.Contains(param.SortBy) ? param.SortBy : sortBy[0]) + sortOrder;
             }
 
             if (param.Fields != null)
             {
-                string[] fieldAllow = { "country_id", "region_id", "country_code", "country_name"};
+                string[] fieldAllow = { "country_id", "region_id", "country_code", "country_name" };
                 List<string> customField = new List<string>();
                 foreach (string field in param.Fields)
                 {
@@ -92,10 +95,42 @@ namespace sync_data.Repositories.ob
 
             string sql = "SELECT " + fields + " FROM " + table;
             sql += where + sort + limit;
-
-            return _connection.Query<CountryObModel>(sql, param);
+            var res = await _connection.QueryAsync<dynamic>(sql, param);
+            if (res == null)
+            {
+                return null;
+            }
+            var data = res.Select(p => new CountryObModel
+            {
+                RegionId = p.region_id ?? null,
+                CountryId = p.country_id ?? null,
+                CountryCode = p.country_code ?? null,
+                CountryName = p.country_name ?? null
+            });
+            return data;
         }
 
+        public async Task<CountryObModel?> GetById(long id)
+        {
+            if (id > 0)
+            {
+                string sql = "SELECT * FROM " + table + " WHERE country_id = @Id";
+                var res = await _connection.QueryFirstOrDefaultAsync<dynamic>(sql, new { Id = id });
+                if (res == null)
+                {
+                    return null;
+                }
+                var data = new CountryObModel()
+                {
+                    CountryId = res.country_id ?? null,
+                    CountryName = res.country_name ?? null,
+                    CountryCode = res.country_code ?? null,
+                    RegionId = res.region_id ?? null,
+                };
+                return data;
+            }
+            return null;
+        }
         public void Insert(CountryObModel Country)
         {
             List<string> column = new List<string>();
@@ -120,7 +155,7 @@ namespace sync_data.Repositories.ob
                 column.Add("region_id");
                 columnData.Add("@RegionId");
             }
-            
+
             string query = "INSERT INTO " + table + "(" + string.Join(", ", column) + ") VALUES (" + string.Join(", ", columnData) + ")";
             _connection.Execute(query, Country);
         }
@@ -146,13 +181,13 @@ namespace sync_data.Repositories.ob
                 {
                     column.Add("region_id");
                 }
-                
-                using (var writer = _connection.BeginBinaryImport("COPY " + table + " (" + string.Join(", ", column)+ ") FROM STDIN (FORMAT BINARY)"))
+
+                using (var writer = _connection.BeginBinaryImport("COPY " + table + " (" + string.Join(", ", column) + ") FROM STDIN (FORMAT BINARY)"))
                 {
                     foreach (var item in countries)
                     {
                         writer.StartRow();
-                        if(item.CountryId != null)
+                        if (item.CountryId != null)
                         {
                             writer.Write(item.CountryId, NpgsqlTypes.NpgsqlDbType.Integer);
                         }
@@ -181,7 +216,7 @@ namespace sync_data.Repositories.ob
             List<string> column = new List<string>();
             List<string> dataSet = new List<string>();
             List<string> dataUpdate = new List<string>();
-            
+
             if (Country.CountryCode != null)
             {
                 column.Add("country_code");
@@ -206,16 +241,16 @@ namespace sync_data.Repositories.ob
                 dataSet.Add("@RegionId");
                 dataUpdate.Add("region_id = @RegionId");
             }
-            
+
             query += string.Join(", ", column) + ") VALUES (" + string.Join(", ", dataSet) + ") ON CONFLICT (country_id) DO UPDATE SET " + string.Join(", ", dataUpdate);
             _connection.Execute(query, Country);
         }
 
-        public void Update(int id, CountryObModel Country)
+        public void Update(long id, CountryObModel Country)
         {
             string query = "UPDATE " + table + " SET ";
             List<string> dataSet = new List<string>();
-            if(Country.CountryId != null)
+            if (Country.CountryId != null)
             {
                 dataSet.Add("country_id = @NameCountryId");
             }
@@ -235,10 +270,13 @@ namespace sync_data.Repositories.ob
             _connection.Execute(query, Country);
         }
 
-        public void Delete(int id)
+        public void Delete(long id)
         {
-            string query = "DELETE FROM " + table + " WHERE country_id = @CountryId";
-            _connection.Execute(query, new { CountryId = id } );
+            if (id > 0)
+            {
+                string query = "DELETE FROM " + table + " WHERE country_id = @CountryId";
+                _connection.Execute(query, new { CountryId = id });
+            }
         }
     }
 }
